+++ src/JestProcessManagement/index.ts
@@ -1,0 +1,2 @@
+export { JestProcess } from './JestProcess'
+export { JestProcessManager } from './JestProcessManager'